requirements.txt
discord.py
requests
python-dotenv
captcha
websockets
qrcode[pil]